Don't install wxPython in the virtualenv
For some reason, pip starts to grab/install wxPython 2.9 when wxPython
2.8 is already installed. If I use "pip install wxpython==2.8" instead,
I just get an error that a matching version can't be found. Since we
create the virtualenv using --system-site-packages, wxPython will get
loaded properly if it's available in the main python install whether or
not we install it within the virtualenv.