INSTALL-xubuntu-12.10: Restored ability to paste multiple apt-get install lines.
When one pastes many command lines and one of those commands has 'apt-get
install -y <packagename>', if that apt-get install asks y/n, the -y will be
used to answer yes to that question, and then all subsequent pasted input
(command lines) will be gobbled up and ignored.

For example, consider pasting the following block of commands (pasted without
the indentation).

    start_time=`date`
    sudo apt-get update
    # sudo apt-get upgrade
    sudo apt-get install -y git
    sudo apt-get install -y python-virtualenv
    sudo apt-get install -y python-dev
    sudo apt-get install -y libfreetype6-dev ;# for matplotlib
    sudo apt-get install -y libpng12-dev ;# for matplotlib
    sudo apt-get install -y libatlas-base-dev ;# for scipy
    sudo apt-get install -y gfortran ;# for scipy
    sudo apt-get install -y python-vtk ;# for mayavi
    sudo apt-get install -y python-wxgtk2.8 ;# this eliminated need for pip install wxPython

    # git clone https://github.com/brandon-rhodes/astronomy-notebooks/
    git clone https://github.com/james-prior/astronomy-notebooks/
    cd astronomy-notebooks
    git checkout v1.1
    # ./setup.sh
    cp -p setup.sh xubuntu-12.10-setup.sh
    sed -e '/vtk/d' <setup.sh >xubuntu-12.10-setup.sh
    ./xubuntu-12.10-setup.sh
    source venv/bin/activate
    # mathjax is not necessary, but cool
    echo 'from IPython.external import mathjax; mathjax.install_mathjax();' \
    | ipython
    echo $start_time
    date
    # Takes about 47 minutes on VM with one core of 2 GHz Intel T2500.
    deactivate

Start with an installation of Xubuntu 12.10 that does not yet have the git or
python-virtualenv packages installed. The first time one pastes the block of
commands, the git package is installed, and all subsequent commands are ignored
as input to 'sudo apt-get install -y git'. The second time one pastes, git is
already installed, so the 'sudo apt-get install -y git' does not ask yes or no
and so does not gobble the rest of the pasted input, so the next command 'sudo
apt-get install -y python-virtualenv' executes, gets a yes from the -y option
and then gobbles the rest of the pasted input. The end of history output is:

   74  sudo true
   75  start_time=`date`
   76  sudo apt-get update
   77  # sudo apt-get upgrade
   78  sudo apt-get install -y git
   79  start_time=`date`
   80  sudo apt-get update
   81  # sudo apt-get upgrade
   82  sudo apt-get install -y git
   83  sudo apt-get install -y python-virtualenv
   84  history

Command #74 is where I take care of entering the password for sudo before I
paste. Commands #75-#78 are executed the first time I paste. Commands #79-83
are executed the second time I paste.

To paste multiple apt-get install commands, the solution is the ugly 'yes |
sudo apt-get install <packagename>'. It gets all its input from the pipe, so it
does not goggle up the pasted input.