Fix undefined name "s"
TODO: must learn how to run the IPython Notebook Viewer locally so that
I can debug all of this locally without having to do round-trips through
git commits.  OH WAIT, I guess I could copy the notebook to a throw-away
URL and view it outside of GitHub itself.  Okay, I will do that from now
on.  But for the moment: I'm doing this commit and pushing to GitHub.