Continue trying the problem with simple geometry
Which won’t work, because the epicycle is at a different distance at
different times — and to find out how far away, we can’t simply treat
the deferent as providing a direction in the sky, but a position at a
specific distance.  So it’s time to upgrade this to Cartesian!