Try controlling rotation with splines; dead end
The rotation is not yet timed correctly, but it turns out this won’t
work: a deferent whose rotation is non-linear will cause the epicycle to
also have non-linear rotation, unless a reverse rotation is applied that
follows the same splines.  And who wants to do that?  I’ll try again,
this time modeling the deferent as an SVG path to travel rather than as
a rotation.